DRILL-6138: Move RecordBatchSizer to org.apache.drill.exec.record package
